+++ libraries/muxer/src/test/java/androidx/media3/muxer/Mp4MuxerEndToEndTest.java
@@ -41,7 +41,6 @@ import androidx.media3.test.utils.FakeExtractorOutput;
 import androidx.media3.test.utils.TestUtil;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-import java.io.FileOutputStream;
 import java.nio.ByteBuffer;
 import org.junit.Rule;
 import org.junit.Test;
@@ -60,7 +59,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void writeMp4File_withSampleAndMetadata_matchedExpectedBoxStructure() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     Pair<ByteBuffer, BufferInfo> sampleAndSampleInfo =
         getFakeSampleAndSampleInfo(/* presentationTimeUs= */ 0L);
     byte[] xmpBytes = TestUtil.getByteArray(context, XMP_SAMPLE_DATA);
@@ -101,7 +102,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_addTrackAndMetadataButNoSamples_createsEmptyFile() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       mp4Muxer.addTrack(/* sortKey= */ 0, FAKE_VIDEO_FORMAT);
@@ -122,7 +125,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createAv1Mp4File_withoutCsd_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       mp4Muxer.addMetadataEntry(
@@ -151,7 +156,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_muxerNotClosed_createsPartiallyWrittenValidFile() throws Exception {
     String outputPath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputPath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputPath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -177,7 +184,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_withSameTracksOffset_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -221,7 +230,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_withDifferentTracksOffset_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -261,7 +272,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_withNegativeTracksOffset_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -298,7 +311,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_withOutOfOrderBframes_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -334,7 +349,9 @@ public class Mp4MuxerEndToEndTest {
   public void createMp4File_withOutOfOrderBframesLargePresentationTimestamps_matchesExpected()
       throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -382,7 +399,9 @@ public class Mp4MuxerEndToEndTest {
   @Test
   public void createMp4File_withOneTrackEmpty_doesNotWriteEmptyTrack() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer mp4Muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     mp4Muxer.addMetadataEntry(
         new Mp4TimestampData(
             /* creationTimestampSeconds= */ 100_000_000L,
@@ -415,7 +434,9 @@ public class Mp4MuxerEndToEndTest {
   public void writeMp4File_withLargeNumberOfSamples_writesMoovBoxAtTheEndAndFreeBoxAtStart()
       throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
     try {
       muxer.addMetadataEntry(
           new Mp4TimestampData(
@@ -445,7 +466,7 @@ public class Mp4MuxerEndToEndTest {
       throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setAttemptStreamableOutputEnabled(false)
             .build();
     try {
@@ -480,7 +501,7 @@ public class Mp4MuxerEndToEndTest {
     assertThrows(
         IllegalArgumentException.class,
         () ->
-            new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+            new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
                 .setOutputFileFormat(Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION)
                 .build());
   }
@@ -492,7 +513,7 @@ public class Mp4MuxerEndToEndTest {
     String outputFilePath = temporaryFolder.newFile().getPath();
     String cacheFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setOutputFileFormat(Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION)
             .setMp4AtFileParameters(
                 new Mp4Muxer.Mp4AtFileParameters(
@@ -539,7 +560,9 @@ public class Mp4MuxerEndToEndTest {
   public void writeMp4File_withDefaultFileFormatAndAuxiliaryTracks_doesNotWriteAxteBox()
       throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -584,7 +607,7 @@ public class Mp4MuxerEndToEndTest {
     String outputFilePath = temporaryFolder.newFile().getPath();
     String cacheFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setOutputFileFormat(Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION)
             .setMp4AtFileParameters(
                 new Mp4Muxer.Mp4AtFileParameters(
@@ -635,7 +658,7 @@ public class Mp4MuxerEndToEndTest {
     String outputFilePath = temporaryFolder.newFile().getPath();
     String cacheFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setOutputFileFormat(Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION)
             .setMp4AtFileParameters(
                 new Mp4Muxer.Mp4AtFileParameters(
@@ -686,7 +709,7 @@ public class Mp4MuxerEndToEndTest {
           throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setOutputFileFormat(Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION)
             .setMp4AtFileParameters(
                 new Mp4Muxer.Mp4AtFileParameters(
@@ -737,7 +760,7 @@ public class Mp4MuxerEndToEndTest {
           throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setOutputFileFormat(Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION)
             .setMp4AtFileParameters(
                 new Mp4Muxer.Mp4AtFileParameters(
@@ -789,7 +812,7 @@ public class Mp4MuxerEndToEndTest {
           throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer mp4Muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setLastSampleDurationBehavior(
                 LAST_SAMPLE_DURATION_BEHAVIOR_SET_FROM_END_OF_STREAM_BUFFER_OR_DUPLICATE_PREVIOUS)
             .build();
@@ -836,7 +859,7 @@ public class Mp4MuxerEndToEndTest {
           throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer mp4Muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .setLastSampleDurationBehavior(
                 LAST_SAMPLE_DURATION_BEHAVIOR_SET_FROM_END_OF_STREAM_BUFFER_OR_DUPLICATE_PREVIOUS)
             .build();
@@ -874,7 +897,7 @@ public class Mp4MuxerEndToEndTest {
     String outputPath = temporaryFolder.newFile().getPath();
 
     Mp4Muxer mp4Muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputPath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputPath)))
             .setSampleBatchingEnabled(false)
             .build();
     mp4Muxer.addMetadataEntry(
@@ -902,7 +925,7 @@ public class Mp4MuxerEndToEndTest {
     String outputPath = temporaryFolder.newFile().getPath();
 
     Mp4Muxer mp4Muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputPath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputPath)))
             .setSampleBatchingEnabled(false)
             .setAttemptStreamableOutputEnabled(false)
             .build();
@@ -932,7 +955,7 @@ public class Mp4MuxerEndToEndTest {
           throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
     Mp4Muxer muxer =
-        new Mp4Muxer.Builder(new FileOutputStream(outputFilePath))
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
             .experimentalSetFreeSpaceAfterFileTypeBox(500)
             .setAttemptStreamableOutputEnabled(false)
             .build();