+++ libraries/muxer/src/test/java/androidx/media3/muxer/Mp4MuxerMetadataTest.java
@@ -37,7 +37,6 @@ import androidx.media3.test.utils.FakeExtractorOutput;
 import androidx.media3.test.utils.TestUtil;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-import java.io.FileOutputStream;
 import java.nio.ByteBuffer;
 import org.junit.Rule;
 import org.junit.Test;
@@ -56,7 +55,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_orientationNotSet_setsOrientationTo0() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -82,7 +83,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_setOrientationTo90_setsOrientationTo90() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -110,7 +113,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_setOrientationTo180_setsOrientationTo180() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -138,7 +143,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_setOrientationTo270_setsOrientationTo270() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -166,7 +173,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_setLocation_setsSameLocation() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -193,7 +202,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_locationNotSet_setsLocationToNull() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -219,7 +230,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_setFrameRate_setsSameFrameRate() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -249,7 +262,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_addStringMetadata_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -278,7 +293,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_addManyLargeStringMetadata_doesNotThrow() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     String metadataKey = "SomeStringKey";
     byte[] metadataValue = Util.getUtf8Bytes("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA");
@@ -298,7 +315,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_addFloatMetadata_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(
@@ -328,7 +347,9 @@ public class Mp4MuxerMetadataTest {
   @Test
   public void writeMp4File_addXmp_matchesExpected() throws Exception {
     String outputFilePath = temporaryFolder.newFile().getPath();
-    Mp4Muxer muxer = new Mp4Muxer.Builder(new FileOutputStream(outputFilePath)).build();
+    Mp4Muxer muxer =
+        new Mp4Muxer.Builder(new MuxerOutputFactoryImpl(SeekableMuxerOutput.of(outputFilePath)))
+            .build();
 
     try {
       muxer.addMetadataEntry(